terminado
proyecto terminado

+++ InventarioVentas/Models/Venta.cs
@@ -5,9 +5,9 @@
         public int IdVenta { get; set; }
         public string NombreProducto { get; set; }
         public DateTime Fecha { get; set; }
-        public double PrecioProducto { get; set; }
+        public decimal PrecioProducto { get; set; }
         public int Cantidad { get; set; }
-        public double Total { get; set; }
+        public decimal Total { get; set; }
 
         //relacion
         public int IdProducto { get; set; }
